+++ src/modules/project/components/index.ts
@@ -1,2 +1,0 @@
-export * from './project-badge-list'
-export * from './project-card-list'