+++ src/modules/project/typings/index.ts
@@ -1,5 +1,0 @@
-export type ProjectMinimal = {
-  slug: string
-  name: string
-  avatar: string
-}