+++ src/modules/tenant/components/index.ts
@@ -1,2 +1,0 @@
-export * from './team-switch'
-export * from './team-switch-ui'