+++ src/modules/workspaces/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './protect-workspace'
+export * from './workspace-user-list'