+++ src/modules/workspaces/typings/index.ts
@@ -1,0 +1,5 @@
+export type WorkspaceMinimal = {
+  slug: string
+  name: string
+  avatar: string
+}