+++ src/typings/general.type.ts
@@ -1,3 +1,3 @@
-export type TenantParams = {
-  params: Promise<{ tenant: string }>
+export type WorkspaceParams = {
+  params: Promise<{ workspace: string }>
 }